// client/src/generated/accounts/userAccount.ts
export type UserAccount = {
  discriminator: number;
  x25519Pubkey: Array<number>;
  threads: Array<ThreadEntry>;
};